[IDP-2366] Handler suppress warning too many parameters into constructors preAuthPayment and authPayment
