+++ src/main/java/it/gov/pagopa/payment/service/payment/barcode/BarCodeAuthPaymentServiceImpl.java
@@ -3,24 +3,18 @@ package it.gov.pagopa.payment.service.payment.barcode;
 import it.gov.pagopa.common.utils.CommonUtilities;
 import it.gov.pagopa.payment.connector.rest.merchant.MerchantConnector;
 import it.gov.pagopa.payment.connector.rest.merchant.dto.MerchantDetailDTO;
-import it.gov.pagopa.payment.connector.rest.reward.RewardCalculatorConnector;
-import it.gov.pagopa.payment.connector.rest.wallet.WalletConnector;
 import it.gov.pagopa.payment.constants.PaymentConstants;
 import it.gov.pagopa.payment.constants.PaymentConstants.ExceptionCode;
 import it.gov.pagopa.payment.dto.AuthPaymentDTO;
 import it.gov.pagopa.payment.dto.barcode.AuthBarCodePaymentDTO;
 import it.gov.pagopa.payment.exception.custom.TransactionInvalidException;
 import it.gov.pagopa.payment.model.TransactionInProgress;
-import it.gov.pagopa.payment.repository.TransactionInProgressRepository;
-import it.gov.pagopa.payment.service.messagescheduler.AuthorizationTimeoutSchedulerServiceImpl;
 import it.gov.pagopa.payment.service.payment.barcode.expired.BarCodeAuthorizationExpiredService;
 import it.gov.pagopa.payment.service.payment.common.CommonAuthServiceImpl;
-import it.gov.pagopa.payment.service.payment.common.CommonPreAuthServiceImpl;
 import it.gov.pagopa.payment.utils.AuditUtilities;
 import it.gov.pagopa.payment.utils.CommonPaymentUtilities;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.tuple.Pair;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import java.util.Collections;
@@ -28,22 +22,20 @@ import java.util.List;
 
 @Slf4j
 @Service
-public class BarCodeAuthPaymentServiceImpl extends CommonAuthServiceImpl implements BarCodeAuthPaymentService {
+public class BarCodeAuthPaymentServiceImpl implements BarCodeAuthPaymentService {
 
     private final BarCodeAuthorizationExpiredService barCodeAuthorizationExpiredService;
     private final MerchantConnector merchantConnector;
-    @SuppressWarnings("squid:S00107") // suppressing too many parameters alert
-    public BarCodeAuthPaymentServiceImpl(TransactionInProgressRepository transactionInProgressRepository,
-                                         BarCodeAuthorizationExpiredService barCodeAuthorizationExpiredService,
-                                         RewardCalculatorConnector rewardCalculatorConnector,
-                                         AuditUtilities auditUtilities,
-                                         WalletConnector walletConnector,
+    private final CommonAuthServiceImpl commonAuthService;
+    protected final AuditUtilities auditUtilities;
+    public BarCodeAuthPaymentServiceImpl(BarCodeAuthorizationExpiredService barCodeAuthorizationExpiredService,
                                          MerchantConnector merchantConnector,
-                                         @Qualifier("commonPreAuth")CommonPreAuthServiceImpl commonPreAuthService,
-                                         AuthorizationTimeoutSchedulerServiceImpl timeoutSchedulerService){
-        super(transactionInProgressRepository, rewardCalculatorConnector, auditUtilities, walletConnector, commonPreAuthService, timeoutSchedulerService);
+                                         CommonAuthServiceImpl commonAuthService,
+                                         AuditUtilities auditUtilities){
         this.barCodeAuthorizationExpiredService = barCodeAuthorizationExpiredService;
         this.merchantConnector = merchantConnector;
+        this.commonAuthService = commonAuthService;
+        this.auditUtilities = auditUtilities;
     }
 
     @Override
@@ -55,17 +47,17 @@ public class BarCodeAuthPaymentServiceImpl extends CommonAuthServiceImpl impleme
             }
 
             TransactionInProgress trx = barCodeAuthorizationExpiredService.findByTrxCodeAndAuthorizationNotExpired(trxCode.toLowerCase());
-            checkAuth(trxCode, trx);
+            commonAuthService.checkAuth(trxCode, trx);
 
             MerchantDetailDTO merchantDetail = merchantConnector.merchantDetail(merchantId, trx.getInitiativeId());
 
-            checkWalletStatus(trx.getInitiativeId(), trx.getUserId());
+            commonAuthService.checkWalletStatus(trx.getInitiativeId(), trx.getUserId());
 
             setTrxFields(merchantId, authBarCodePaymentDTO, trx, merchantDetail, acquirerId);
 
-            checkTrxStatusToInvokePreAuth(trx);
+            commonAuthService.checkTrxStatusToInvokePreAuth(trx);
 
-            AuthPaymentDTO authPaymentDTO = invokeRuleEngine(trx);
+            AuthPaymentDTO authPaymentDTO = commonAuthService.invokeRuleEngine(trx);
 
             logAuthorizedPayment(authPaymentDTO.getInitiativeId(), authPaymentDTO.getId(), trxCode, merchantId,authPaymentDTO.getReward(), authPaymentDTO.getRejectionReasons());
             authPaymentDTO.setResidualBudget(CommonPaymentUtilities.calculateResidualBudget(authPaymentDTO.getRewards()));
@@ -80,13 +72,11 @@ public class BarCodeAuthPaymentServiceImpl extends CommonAuthServiceImpl impleme
         }
     }
 
-    @Override
-    protected void logAuthorizedPayment(String initiativeId, String id, String trxCode, String merchantId,Long reward, List<String> rejectionReasons){
+    private void logAuthorizedPayment(String initiativeId, String id, String trxCode, String merchantId,Long reward, List<String> rejectionReasons){
         auditUtilities.logBarCodeAuthorizedPayment(initiativeId, id, trxCode, merchantId, reward, rejectionReasons);
     }
 
-    @Override
-    protected void logErrorAuthorizedPayment(String trxCode, String merchantId){
+    private void logErrorAuthorizedPayment(String trxCode, String merchantId){
         auditUtilities.logBarCodeErrorAuthorizedPayment(trxCode, merchantId);
     }
