+++ src/main/java/it/gov/pagopa/payment/service/payment/common/CommonPreAuthServiceImpl.java
@@ -102,7 +102,7 @@ public class CommonPreAuthServiceImpl{
     }
   }
 
-  protected void checkPreAuth(String userId, TransactionInProgress trx) {
+  public void checkPreAuth(String userId, TransactionInProgress trx) {
     String walletStatus = walletConnector.getWallet(trx.getInitiativeId(), userId).getStatus();
     if (PaymentConstants.WALLET_STATUS_SUSPENDED.equals(walletStatus)){
       throw new UserSuspendedException("The user has been suspended for initiative [%s]".formatted(trx.getInitiativeId()));
@@ -129,7 +129,7 @@ public class CommonPreAuthServiceImpl{
     }
   }
 
-  protected void auditLogRelateUser(TransactionInProgress trx, String channel){
+  public void auditLogRelateUser(TransactionInProgress trx, String channel){
     auditUtilities.logRelatedUserToTransaction(trx.getInitiativeId(), trx.getId(), trx.getTrxCode(), trx.getUserId(), channel);
   }
 }
