+++ src/main/java/it/gov/pagopa/payment/service/payment/idpaycode/IdpayCodeAuthPaymentServiceImpl.java
@@ -1,8 +1,6 @@
 package it.gov.pagopa.payment.service.payment.idpaycode;
 
 import it.gov.pagopa.payment.connector.rest.paymentinstrument.PaymentInstrumentConnectorImpl;
-import it.gov.pagopa.payment.connector.rest.reward.RewardCalculatorConnector;
-import it.gov.pagopa.payment.connector.rest.wallet.WalletConnector;
 import it.gov.pagopa.payment.constants.PaymentConstants.ExceptionCode;
 import it.gov.pagopa.payment.dto.AuthPaymentDTO;
 import it.gov.pagopa.payment.dto.PinBlockDTO;
@@ -10,32 +8,21 @@ import it.gov.pagopa.payment.exception.custom.MerchantOrAcquirerNotAllowedExcept
 import it.gov.pagopa.payment.exception.custom.OperationNotAllowedException;
 import it.gov.pagopa.payment.exception.custom.TransactionNotFoundOrExpiredException;
 import it.gov.pagopa.payment.model.TransactionInProgress;
-import it.gov.pagopa.payment.repository.TransactionInProgressRepository;
-import it.gov.pagopa.payment.service.messagescheduler.AuthorizationTimeoutSchedulerServiceImpl;
 import it.gov.pagopa.payment.service.payment.common.CommonAuthServiceImpl;
-import it.gov.pagopa.payment.service.payment.common.CommonPreAuthServiceImpl;
 import it.gov.pagopa.payment.service.payment.idpaycode.expired.IdpayCodeAuthorizationExpiredService;
-import it.gov.pagopa.payment.utils.AuditUtilities;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 @Slf4j
 @Service
-public class IdpayCodeAuthPaymentServiceImpl extends CommonAuthServiceImpl implements IdpayCodeAuthPaymentService {
+public class IdpayCodeAuthPaymentServiceImpl implements IdpayCodeAuthPaymentService {
     private final IdpayCodeAuthorizationExpiredService idpayCodeAuthorizationExpiredService;
     private final PaymentInstrumentConnectorImpl paymentInstrumentConnector;
-    @SuppressWarnings("squid:S00107") // suppressing too many parameters alert
-    protected IdpayCodeAuthPaymentServiceImpl(TransactionInProgressRepository transactionInProgressRepository,
-                                              RewardCalculatorConnector rewardCalculatorConnector,
-                                              AuditUtilities auditUtilities,
-                                              WalletConnector walletConnector,
-                                              IdpayCodeAuthorizationExpiredService idpayCodeAuthorizationExpiredService,
-                                              PaymentInstrumentConnectorImpl paymentInstrumentConnector,
-                                              @Qualifier("commonPreAuth")CommonPreAuthServiceImpl commonPreAuthService,
-                                              AuthorizationTimeoutSchedulerServiceImpl timeoutSchedulerServiceImpl) {
-        super(transactionInProgressRepository, rewardCalculatorConnector, auditUtilities, walletConnector, commonPreAuthService, timeoutSchedulerServiceImpl);
+    private final CommonAuthServiceImpl commonAuthService;
+    protected IdpayCodeAuthPaymentServiceImpl(IdpayCodeAuthorizationExpiredService idpayCodeAuthorizationExpiredService,
+                                              PaymentInstrumentConnectorImpl paymentInstrumentConnector, CommonAuthServiceImpl commonAuthService) {
         this.idpayCodeAuthorizationExpiredService = idpayCodeAuthorizationExpiredService;
         this.paymentInstrumentConnector = paymentInstrumentConnector;
+        this.commonAuthService = commonAuthService;
     }
     @Override
     public AuthPaymentDTO authPayment(String trxId, String merchantId, PinBlockDTO pinBlockBody) {
@@ -55,7 +42,7 @@ public class IdpayCodeAuthPaymentServiceImpl extends CommonAuthServiceImpl imple
         if (!merchantId.equals(trx.getMerchantId())){
             throw new MerchantOrAcquirerNotAllowedException("The merchant with id [%s] associated to the transaction is not equal to the merchant with id [%s]".formatted(trx.getMerchantId(),merchantId));
         }
-        return super.authPayment(trx,trx.getUserId(),trx.getTrxCode());
+        return commonAuthService.authPayment(trx,trx.getUserId(),trx.getTrxCode());
 
     }
 }
