+++ src/main/java/it/gov/pagopa/payment/service/payment/idpaycode/IdpayCodeRelateUserService.java
@@ -1,0 +1,7 @@
+package it.gov.pagopa.payment.service.payment.idpaycode;
+
+import it.gov.pagopa.payment.dto.idpaycode.RelateUserResponse;
+
+public interface IdpayCodeRelateUserService {
+    RelateUserResponse relateUser(String trxId, String fiscalCode);
+}
