+++ src/test/java/it/gov/pagopa/payment/service/payment/barcode/BarCodeAuthPaymentServiceImplTest.java
@@ -1,12 +1,8 @@
 package it.gov.pagopa.payment.service.payment.barcode;
 
-import com.mongodb.client.result.UpdateResult;
 import it.gov.pagopa.common.utils.TestUtils;
 import it.gov.pagopa.payment.connector.rest.merchant.MerchantConnector;
 import it.gov.pagopa.payment.connector.rest.merchant.dto.MerchantDetailDTO;
-import it.gov.pagopa.payment.connector.rest.reward.RewardCalculatorConnector;
-import it.gov.pagopa.payment.connector.rest.wallet.WalletConnector;
-import it.gov.pagopa.payment.connector.rest.wallet.dto.WalletDTO;
 import it.gov.pagopa.payment.constants.PaymentConstants;
 import it.gov.pagopa.payment.dto.AuthPaymentDTO;
 import it.gov.pagopa.payment.dto.Reward;
@@ -17,17 +13,12 @@ import it.gov.pagopa.payment.exception.custom.TransactionInvalidException;
 import it.gov.pagopa.payment.exception.custom.TransactionNotFoundOrExpiredException;
 import it.gov.pagopa.payment.model.TransactionInProgress;
 import it.gov.pagopa.payment.model.counters.RewardCounters;
-import it.gov.pagopa.payment.repository.TransactionInProgressRepository;
-import it.gov.pagopa.payment.service.messagescheduler.AuthorizationTimeoutSchedulerServiceImpl;
 import it.gov.pagopa.payment.service.payment.barcode.expired.BarCodeAuthorizationExpiredService;
 import it.gov.pagopa.payment.service.payment.common.CommonAuthServiceImpl;
-import it.gov.pagopa.payment.service.payment.common.CommonPreAuthServiceImpl;
 import it.gov.pagopa.payment.test.fakers.AuthPaymentDTOFaker;
 import it.gov.pagopa.payment.test.fakers.RewardFaker;
 import it.gov.pagopa.payment.test.fakers.TransactionInProgressFaker;
-import it.gov.pagopa.payment.test.fakers.WalletDTOFaker;
 import it.gov.pagopa.payment.utils.AuditUtilities;
-import it.gov.pagopa.payment.utils.CommonPaymentUtilities;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -37,26 +28,16 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 class BarCodeAuthPaymentServiceImplTest {
 
-    @Mock
-    private TransactionInProgressRepository repositoryMock;
     @Mock private BarCodeAuthorizationExpiredService barCodeAuthorizationExpiredServiceMock;
-    @Mock private RewardCalculatorConnector rewardCalculatorConnectorMock;
     @Mock private AuditUtilities auditUtilitiesMock;
-    @Mock private WalletConnector walletConnectorMock;
     @Mock private MerchantConnector merchantConnector;
-    @Mock private CommonPreAuthServiceImpl commonPreAuthServiceMock;
-
-    @Mock private AuthorizationTimeoutSchedulerServiceImpl timeoutSchedulerServiceMock;
     @Mock private CommonAuthServiceImpl commonAuthServiceMock;
     private static final String USER_ID = "USERID1";
     private static final String MERCHANT_ID = "MERCHANT_ID";
@@ -75,14 +56,10 @@ class BarCodeAuthPaymentServiceImplTest {
     @BeforeEach
     void setup(){
         barCodeAuthPaymentService = new BarCodeAuthPaymentServiceImpl(
-                repositoryMock,
                 barCodeAuthorizationExpiredServiceMock,
-                rewardCalculatorConnectorMock,
-                auditUtilitiesMock,
-                walletConnectorMock,
                 merchantConnector,
-                commonPreAuthServiceMock,
-                timeoutSchedulerServiceMock);
+                commonAuthServiceMock,
+                auditUtilitiesMock);
     }
 
     @Test
@@ -98,34 +75,21 @@ class BarCodeAuthPaymentServiceImplTest {
         Reward reward = RewardFaker.mockInstance(1);
         reward.setCounters(new RewardCounters());
 
-        WalletDTO walletDTO = WalletDTOFaker.mockInstance(1, "REFUNDABLE");
-        when(walletConnectorMock.getWallet(any(), any())).thenReturn(walletDTO);
 
         when(barCodeAuthorizationExpiredServiceMock.findByTrxCodeAndAuthorizationNotExpired(transaction.getTrxCode()))
                 .thenReturn(transaction);
 
         when(merchantConnector.merchantDetail(MERCHANT_ID, transaction.getInitiativeId()))
                 .thenReturn(MerchantDetailDTO.builder().build());
-
-        when(timeoutSchedulerServiceMock.scheduleMessage(transaction.getId())).thenReturn(1L);
-
-        when(rewardCalculatorConnectorMock.authorizePayment(transaction)).thenReturn(authPaymentDTO);
-
-        when(repositoryMock.updateTrxAuthorized(transaction, authPaymentDTO,
-                CommonPaymentUtilities.getInitiativeRejectionReason(transaction.getInitiativeId(), List.of())))
-                .thenReturn(UpdateResult.acknowledged(1, 1L, null));
-
-        doNothing().when(timeoutSchedulerServiceMock).cancelScheduledMessage(1L);
-
+        when(commonAuthServiceMock.invokeRuleEngine(transaction))
+                .thenReturn(authPaymentDTO);
         // When
         AuthPaymentDTO result = barCodeAuthPaymentService.authPayment(TRX_CODE1, AUTH_BAR_CODE_PAYMENT_DTO, MERCHANT_ID, ACQUIRER_ID);
 
         // Then
-        verify(barCodeAuthorizationExpiredServiceMock).findByTrxCodeAndAuthorizationNotExpired(TRX_CODE1);
-        verify(walletConnectorMock, times(1)).getWallet(transaction.getInitiativeId(), USER_ID);
-        verify(timeoutSchedulerServiceMock, times(1)).scheduleMessage(transaction.getId());
-        verify(timeoutSchedulerServiceMock, times(1)).cancelScheduledMessage(1L);
+        assertNotNull(result);
         assertEquals(authPaymentDTO, result);
+        verify(barCodeAuthorizationExpiredServiceMock).findByTrxCodeAndAuthorizationNotExpired(TRX_CODE1);
         TestUtils.checkNotNullFields(result, "rejectionReasons","splitPayment",
                 "residualAmountCents");
         assertEquals(transaction.getTrxCode(), result.getTrxCode());
@@ -153,6 +117,9 @@ class BarCodeAuthPaymentServiceImplTest {
         // Given
         when(barCodeAuthorizationExpiredServiceMock.findByTrxCodeAndAuthorizationNotExpired(TRX_CODE1)).thenReturn(null);
 
+
+        doThrow(new TransactionNotFoundOrExpiredException("DUMMY_EXCEPTION"))
+                .when(commonAuthServiceMock).checkAuth(eq(TRX_CODE1), any());
         // When
         TransactionNotFoundOrExpiredException result =
                 assertThrows(TransactionNotFoundOrExpiredException.class, () -> barCodeAuthPaymentService.authPayment(TRX_CODE1, AUTH_BAR_CODE_PAYMENT_DTO, MERCHANT_ID, ACQUIRER_ID));
@@ -171,15 +138,17 @@ class BarCodeAuthPaymentServiceImplTest {
         AuthPaymentDTO authPaymentDTO = AuthPaymentDTOFaker.mockInstance(1, transaction);
         authPaymentDTO.setStatus(SyncTrxStatus.REWARDED);
 
-        Reward reward = RewardFaker.mockInstance(1);
-        reward.setCounters(new RewardCounters());
+        when(barCodeAuthorizationExpiredServiceMock.findByTrxCodeAndAuthorizationNotExpired(transaction.getTrxCode()))
+                .thenReturn(transaction);
 
+        when(merchantConnector.merchantDetail(MERCHANT_ID, transaction.getInitiativeId()))
+                .thenReturn(MerchantDetailDTO.builder().build());
 
-        when(rewardCalculatorConnectorMock.authorizePayment(transaction)).thenThrow(new TooManyRequestsException("Too many request on the ms reward",true,null));
-
+        when(commonAuthServiceMock.invokeRuleEngine(transaction))
+                .thenThrow(new TooManyRequestsException("Too many request on the ms reward",true,null));
 
         TooManyRequestsException result = Assertions.assertThrows(TooManyRequestsException.class,
-                () -> rewardCalculatorConnectorMock.authorizePayment(transaction));
+                () -> barCodeAuthPaymentService.authPayment(TRX_CODE1, AUTH_BAR_CODE_PAYMENT_DTO, MERCHANT_ID, ACQUIRER_ID));
 
         // Then
         Assertions.assertEquals(PaymentConstants.ExceptionCode.TOO_MANY_REQUESTS, result.getCode());
